First try for screen sharing, on Chrome.
The Screen Sharing functionality requires a Chrome Extension
to manage it. This first version comes with a locally installed
extension.

We cannot have both audio + screenshare in a single stream, so we
call getUserMedia a first time to get access to the Screen, then
a second time to get access to the microphone, and we finally
add the audioTrack from the microphone stream to the video
stream that matches with the screen.

More details here :
https://stackoverflow.com/questions/19382444/is-it-possible-broadcast-audio-with-screensharing-with-webrtc